vari fix

+++ src/main/java/it/polimi/ingsw/GC_15/Dice.java
@@ -21,7 +21,7 @@ public class Dice {
 	
 	public void setValue() {
 		Random r = new Random();
-		value= (int) (r.nextInt(6) + 1);
+		value= r.nextInt(6) + 1;
 	}
 	
 	public void setValue(int value){
